+++ src/main/java/io/github/connortron110/scplockdown/client/renderer/ItemModelRenderer.java
@@ -1,85 +1,95 @@
 package io.github.connortron110.scplockdown.client.renderer;
 
-/*import com.mojang.blaze3d.vertex.PoseStack;
-import io.github.connortron110.scplockdown.SCPLockdown;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import io.github.connortron110.scplockdown.client.SCPLayerDefinitions;
+import io.github.connortron110.scplockdown.client.models.SCP035MaskModel;
+import io.github.connortron110.scplockdown.client.renderer.blockentity.SCP035CaseRenderer;
+import io.github.connortron110.scplockdown.level.blockentity.BlastDoorBlockEntity;
 import io.github.connortron110.scplockdown.level.blockentity.LockerBlockEntity;
+import io.github.connortron110.scplockdown.level.blockentity.SlidingDoorBlockEntity;
+import io.github.connortron110.scplockdown.level.blocks.BlastDoorBlock;
+import io.github.connortron110.scplockdown.level.blocks.LockerBlock;
+import io.github.connortron110.scplockdown.level.blocks.SlidingDoorBlock;
 import io.github.connortron110.scplockdown.level.items.SCP035MaskItem;
-import io.github.connortron110.scplockdown.level.tileentity.LockerTileEntity;
+import io.github.connortron110.scplockdown.registration.SCPBlocks;
+import io.github.connortron110.scplockdown.utils.ClientUtils;
 import net.minecraft.client.model.geom.EntityModelSet;
-import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderDispatcher;
+import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.client.resources.model.Material;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.client.resources.model.ModelBakery;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraftforge.client.ForgeHooksClient;
-import net.minecraftforge.common.util.Lazy;
 
-public class ItemModelRenderer extends BlockEntityWithoutLevelRenderer {
-    public static final ItemModelRenderer INSTANCE = new ItemModelRenderer();
+/**
+ * Renders custom models into the world (inventory, in hand etc).
+ * To avoid clashing with other mods, we use {@link io.github.connortron110.scplockdown.mixin.client.ClientMixinBlockEntityWithoutLevelRenderer} to "hook" into the main rendering class to then redirect calls to here once all vanilla / other mods are done with their stuff.
+ * All functions are derived from {@link net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer}.
+ */
+public class ItemModelRenderer {
 
-    public static final Material SCP035MATERIAL = ForgeHooksClient.getBlockMaterial(ResourceLocation.fromNamespaceAndPath(SCPLockdown.MOD_ID, "entity/scp035")); //FIXME Texture needs to be square
+	private static final LockerBlockEntity LOCKER = new LockerBlockEntity(BlockPos.ZERO, SCPBlocks.LOCKER.get().defaultBlockState());
+	private static final SlidingDoorBlockEntity SLIDING_DOOR = new SlidingDoorBlockEntity(BlockPos.ZERO, SCPBlocks.SLIDING_DOOR.get().defaultBlockState());
+	private static final SlidingDoorBlockEntity CONTAINMENT_DOOR = new SlidingDoorBlockEntity(BlockPos.ZERO, SCPBlocks.CONTAINMENT_DOOR.get().defaultBlockState());
+	private static final SlidingDoorBlockEntity MAGNETIZED_DOOR = new SlidingDoorBlockEntity(BlockPos.ZERO, SCPBlocks.MAGNETIZED_DOOR.get().defaultBlockState());
+	private static final BlastDoorBlockEntity BLAST_DOOR = new BlastDoorBlockEntity(BlockPos.ZERO, SCPBlocks.BLAST_DOOR.get().defaultBlockState());
 
-    private final SCP035MaskModel scp035Mask = new SCP035MaskModel();
+	private SCP035MaskModel scp035MaskModel;
 
-    private final Lazy<BlockEntity> locker = Lazy.of(LockerBlockEntity::new);
+	private final BlockEntityRenderDispatcher blockEntityRenderDispatcher;
+	private final EntityModelSet entityModelSet;
 
-    public ItemModelRenderer(BlockEntityRenderDispatcher pBlockEntityRenderDispatcher, EntityModelSet pEntityModelSet) {
-        super(pBlockEntityRenderDispatcher, pEntityModelSet);
-    }
+	public ItemModelRenderer(BlockEntityRenderDispatcher pBlockEntityRenderDispatcher, EntityModelSet pEntityModelSet) {
+		this.blockEntityRenderDispatcher = pBlockEntityRenderDispatcher;
+		this.entityModelSet = pEntityModelSet;
+	}
 
-    @Override
-    public void renderByItem(ItemStack pStack, ItemDisplayContext pDisplayContext, PoseStack pPoseStack, MultiBufferSource pBuffer, int pPackedLight, int pPackedOverlay) {
-        Item item = pStack.getItem();
-        pPoseStack.pushPose();
-        pPoseStack.translate(0.5D, 0.5D, 0.5D);
-        pPoseStack.scale(0.75F, 0.75F, 0.75F);
-        pPoseStack.mulPose(new Quaternion(Vector3f.YP, 180, true));
-        if (item instanceof SCP035MaskItem) render035Mask(pStack, pDisplayContext, pPoseStack, pBuffer, pPackedLight, pPackedOverlay);
-        else if (item instanceof BlockItem) {
-            TileEntityRendererDispatcher.instance.renderItem(locker.get(), pPoseStack, pBuffer, pPackedLight, pPackedOverlay);
-        }
+	public void onResourceManagerReload(ResourceManager pResourceManager) {
+		this.scp035MaskModel = new SCP035MaskModel(this.entityModelSet.bakeLayer(SCPLayerDefinitions.SCP035_MASK));
+	}
 
+	public void renderByItem(ItemStack pStack, ItemDisplayContext pDisplayContext, PoseStack pPoseStack, MultiBufferSource pBuffer, int pPackedLight, int pPackedOverlay) {
+		Item item = pStack.getItem();
+		if (item instanceof SCP035MaskItem) {
+			pPoseStack.pushPose();
+			ClientUtils.blockEntityRotation(pPoseStack, 1.0D);
+			VertexConsumer vertexconsumer = SCP035CaseRenderer.SCP035_MATERIAL.sprite().wrap(ItemRenderer.getFoilBufferDirect(pBuffer, this.scp035MaskModel.renderType(SCP035CaseRenderer.SCP035_MATERIAL.atlasLocation()), true, pStack.hasFoil()));
+			if (SCP035MaskItem.isComedy(pStack))
+				this.scp035MaskModel.comedy().render(pPoseStack, vertexconsumer, pPackedLight, pPackedOverlay, 1.0F, 1.0F, 1.0F, 1.0F);
+			else
+				this.scp035MaskModel.tragedy().render(pPoseStack, vertexconsumer, pPackedLight, pPackedOverlay, 1.0F, 1.0F, 1.0F, 1.0F);
+			pPoseStack.popPose();
+		}
 
-
-
-        pPoseStack.popPose();
-    }
-
-
-    private void render035Mask(ItemStack pStack, ItemDisplayContext pDisplayContext, PoseStack pPoseStack, MultiBufferSource pBuffer, int pPackedLight, int pPackedOverlay) {
-        if (pDisplayContext == ItemCameraTransforms.TransformType.GUI) {
-            pPoseStack.translate(0.0D, -0.5D, 0.0D);
-            pPoseStack.scale(1.4F, 1.4F, 1.4F);
-            pPoseStack.mulPose(new Quaternion(new Vector3f(-1 ,1, -0.2F), 30, true));
-        }
-
-        if (pDisplayContext == ItemCameraTransforms.TransformType.FIXED) {
-            pPoseStack.translate(0.0D, -0.35D, 0.0D);
-            pPoseStack.mulPose(new Quaternion(Vector3f.YP, 180, true));
-        }
-
-        if (pDisplayContext == ItemCameraTransforms.TransformType.THIRD_PERSON_LEFT_HAND || pDisplayContext == ItemCameraTransforms.TransformType.THIRD_PERSON_RIGHT_HAND) {
-            boolean flag = pDisplayContext == ItemCameraTransforms.TransformType.THIRD_PERSON_RIGHT_HAND;
-            pPoseStack.translate(0.0D, -0.25D, -0.2D);
-            pPoseStack.mulPose(new Quaternion(new Vector3f(0, flag ? 1 : -1, 0), 90, true));
-        }
-
-        if (pDisplayContext == ItemCameraTransforms.TransformType.FIRST_PERSON_LEFT_HAND || pDisplayContext == ItemCameraTransforms.TransformType.FIRST_PERSON_RIGHT_HAND) {
-            boolean flag = pDisplayContext == ItemCameraTransforms.TransformType.FIRST_PERSON_LEFT_HAND;
-            pPoseStack.scale(0.8F, 0.8F, 0.8F);
-            pPoseStack.translate(flag ? 0.25D : -0.25D, 0.0D, 0.0D);
-            pPoseStack.mulPose(new Quaternion(new Vector3f(0, flag ? 1 : -1, 0), 30, true));
-        }
-
-        RenderMaterial rendermaterial = SCP035MATERIAL;
-        IVertexBuilder ivertexbuilder = rendermaterial.sprite().wrap(ItemRenderer.getFoilBufferDirect(pBuffer, this.scp035Mask.renderType(rendermaterial.atlasLocation()), true, pStack.hasFoil()));
-        if (SCP035MaskItem.isComedy(pStack)) this.scp035Mask.comedy();
-        else this.scp035Mask.tragedy();
-        this.scp035Mask.getMaskBase().render(pPoseStack, ivertexbuilder, pPackedLight, pPackedOverlay, 1.0F, 1.0F, 1.0F, 1.0F);
-    }
-}*/
+		if (item instanceof BlockItem blockItem) {
+			if (blockItem.getBlock() instanceof LockerBlock) {
+				pPoseStack.pushPose();
+				ClientUtils.blockEntityRotation(pPoseStack, 1.0D);
+				this.blockEntityRenderDispatcher.renderItem(LOCKER, pPoseStack, pBuffer, pPackedLight, pPackedOverlay);
+				pPoseStack.popPose();
+			} else if (blockItem.getBlock() instanceof SlidingDoorBlock slidingDoorBlock) {
+				pPoseStack.pushPose();
+				ClientUtils.blockEntityRotation(pPoseStack, 1.0D);
+				if (slidingDoorBlock.equals(SCPBlocks.SLIDING_DOOR.get())) {
+					this.blockEntityRenderDispatcher.renderItem(SLIDING_DOOR, pPoseStack, pBuffer, pPackedLight, pPackedOverlay);
+				} else if (slidingDoorBlock.equals(SCPBlocks.CONTAINMENT_DOOR.get())) {
+					this.blockEntityRenderDispatcher.renderItem(CONTAINMENT_DOOR, pPoseStack, pBuffer, pPackedLight, pPackedOverlay);
+				} else if (slidingDoorBlock.equals(SCPBlocks.MAGNETIZED_DOOR.get())) {
+					this.blockEntityRenderDispatcher.renderItem(MAGNETIZED_DOOR, pPoseStack, pBuffer, pPackedLight, pPackedOverlay);
+				}
+				pPoseStack.popPose();
+			} else if (blockItem.getBlock() instanceof BlastDoorBlock) {
+				pPoseStack.pushPose();
+				ClientUtils.blockEntityRotation(pPoseStack, 1.0D);
+				this.blockEntityRenderDispatcher.renderItem(BLAST_DOOR, pPoseStack, pBuffer, pPackedLight, pPackedOverlay);
+				pPoseStack.popPose();
+			}
+		}
+	}
+}